feat(plasmasearchbar): put back QS in the listbox + keeping old version with resultlist
