+++ packages/website/src/searchBar/PlasmaSearchBar.tsx
@@ -1,56 +1,30 @@
 import '@demo-styling/plasmaSearchBar.scss';
 
-import {
-    buildResultList,
-    buildSearchBox,
-    ResultList as HeadlessResultList,
-    SearchBox as HeadlessSearchBox,
-    loadSearchHubActions,
-} from '@coveo/headless';
-import {FunctionComponent, useContext, useEffect, useRef, useState} from 'react';
+import {buildSearchBox, SearchBox as HeadlessSearchBox, SearchBoxOptions} from '@coveo/headless';
+import {FunctionComponent, useContext, useEffect, useState} from 'react';
 import * as React from 'react';
-import {Button, IItemBoxProps, keyCode, ListBox, Svg} from '@coveord/plasma-react';
+import {Button, ListBox, Svg} from '@coveord/plasma-react';
 
 import classNames from 'classnames';
 import {EngineContext} from './engine/EngineContext';
 
 interface ISearchboxProps {
-    id: string;
+    controller: HeadlessSearchBox;
 }
 
-const SearchBoxRenderer: FunctionComponent<{
-    id: string;
-    setSearchHub: (searchHub: string) => void;
-    searchController: HeadlessSearchBox;
-    resultListController: HeadlessResultList;
-}> = (props) => {
-    const {id, setSearchHub, searchController, resultListController} = props;
-    const [stateSearchBox, setStateSearchBox] = useState(searchController.state);
-    const [stateResultList, setStateResultList] = useState(resultListController.state);
-    const previousSearchId = useRef('');
+const SearchBoxRenderer: FunctionComponent<ISearchboxProps> = (props) => {
+    const {controller} = props;
+    const [state, setState] = useState(controller.state);
 
-    useEffect(() => {
-        searchController.subscribe(() => setStateSearchBox(searchController.state));
-        resultListController.subscribe(() => setStateResultList(resultListController.state));
-        setSearchHub('plasmaComponents');
-    }, []);
-
-    const isNewSearchEvent = () => {
-        const currentSearchId = stateResultList.searchResponseId;
-
-        if (currentSearchId !== previousSearchId.current) {
-            previousSearchId.current = currentSearchId;
-            return true;
-        }
-        return false;
-    };
+    useEffect(() => controller.subscribe(() => setState(controller.state)), [controller]);
 
     const ClearButton = () => (
         <button
-            disabled={stateSearchBox.value === ''}
-            className={classNames('clear-button', {'search-not-empty': stateSearchBox.value !== ''})}
+            disabled={state.value === ''}
+            className={classNames('clear-button', {'search-not-empty': state.value !== ''})}
             onClick={() => {
-                searchController.clear();
+                controller.updateText('');
+                controller.submit();
             }}
         >
             <Svg svgName="cross" svgClass="icon" />
@@ -61,70 +35,57 @@ const SearchBoxRenderer: FunctionComponent<{
         <Button
             classes={['search-button']}
             onClick={() => {
-                searchController.submit();
+                controller.submit();
             }}
         >
             <Svg svgName={'search'} className="icon mod-stroke" />
         </Button>
     );
 
-    const ResultList = () => {
-        const navigate = (item: IItemBoxProps) => {
-            window.open(item.value, '_self');
-        };
+    const SearchResultListBox = () => (
+        // const navigate = (item: IItemBoxProps) => {
+        //     window.open(item.value, '_self');
+        // };
 
-        const results: IItemBoxProps[] = stateResultList.results.map((result) => ({
-            value: result.clickUri,
-            displayValue: result.raw.componentname,
-        }));
-
-        return (
-            <ListBox
-                classes={['search-results-container']}
-                isLoading={stateResultList.isLoading}
-                items={results}
-                onOptionClick={navigate}
-            />
-        );
-    };
-
+        <>
+            {(state.suggestions.length > 0 || state.isLoadingSuggestions) && (
+                <ListBox
+                    classes={['search-results-container']}
+                    isLoading={state.isLoadingSuggestions}
+                    items={state.suggestions.map((s) => ({
+                        value: s.rawValue,
+                        displayValue: s.highlightedValue,
+                    }))}
+                    onOptionClick={() => controller.selectSuggestion(state.value)}
+                />
+            )}
+        </>
+    );
     return (
         <div className="plasmaSearchBar">
-            <form
-                autoComplete="off"
-                onSubmit={(e) => {
-                    e.preventDefault();
-                }}
-            >
+            {/* Prevents chrome from providing autocompletions */}
+            <form autoComplete="off" onSubmit={(e) => e.preventDefault()}>
                 <input
-                    id={id}
                     className="search-bar"
-                    type="search"
+                    type="text"
                     placeholder={'Find a component...'}
-                    value={stateSearchBox.value}
-                    onChange={(event) => searchController.updateText(event.target.value)}
-                    onKeyDown={(event) => event.keyCode === keyCode.enter && searchController.submit()}
+                    value={state.value}
+                    onChange={(event) => controller.updateText(event.target.value)}
+                    onKeyDown={(event) => event.key === 'Enter' && controller.submit()}
                 />
                 <ClearButton />
                 <SearchButton />
-                {isNewSearchEvent() && <ResultList />}
+                <SearchResultListBox />
             </form>
         </div>
     );
 };
 
-export const PlasmaSearchBar: FunctionComponent<ISearchboxProps> = ({id}) => {
+const PlasmaSearchBar = () => {
+    const options: SearchBoxOptions = {numberOfSuggestions: 5};
     const engine = useContext(EngineContext);
-    const {setSearchHub} = loadSearchHubActions(engine);
-    const controller = buildSearchBox(engine, {options: {id}});
-    const resultListController = buildResultList(engine, {options: {fieldsToInclude: ['componentname']}});
-
-    return (
-        <SearchBoxRenderer
-            setSearchHub={(searchHub) => engine.dispatch(setSearchHub(searchHub))}
-            searchController={controller}
-            resultListController={resultListController}
-            id={id}
-        />
-    );
+    const controller = buildSearchBox(engine, {options});
+    return <SearchBoxRenderer controller={controller} />;
 };
+
+export default PlasmaSearchBar;
